requirements.txt
numpy >= 1.14
matplotlib >= 2.1
Scipy >= 0.19
scikit-image >= 0.13
pandas >= 0.20
Pillow >= 5.0
tensorboardX >= 1.2
tensorboard >= 1.8.0
tqdm >= 4.0
